chore(types): move overloads from sparqlwrapper to pyi definitions
The change moves typing.overload definitions from sparqlwrapper.py to
an interface module sparqlwrapper.pyi.

Closes #53.
